+++ async_upnp_client/__init__.py
@@ -4,10 +4,10 @@
 from async_upnp_client.advertisement import UpnpAdvertisementListener  # noqa: F401
 from async_upnp_client.client import UpnpAction  # noqa: F401
 from async_upnp_client.client import UpnpDevice  # noqa: F401
-from async_upnp_client.client import UpnpError  # noqa: F401
 from async_upnp_client.client import UpnpRequester  # noqa: F401
 from async_upnp_client.client import UpnpService  # noqa: F401
 from async_upnp_client.client import UpnpStateVariable  # noqa: F401
-from async_upnp_client.client import UpnpValueError  # noqa: F401
 from async_upnp_client.client_factory import UpnpFactory  # noqa: F401
+from async_upnp_client.exceptions import UpnpError  # noqa: F401
+from async_upnp_client.exceptions import UpnpValueError  # noqa: F401
 from async_upnp_client.event_handler import UpnpEventHandler  # noqa: F401