+++ tests/upnp_test_requester.py
@@ -3,13 +3,14 @@
 
 import asyncio
 import os.path
-from copy import copy
-from typing import Mapping, Optional, Tuple, Union
+from collections import deque
+from copy import deepcopy
+from typing import Any, Deque, Mapping, MutableMapping, Optional, Tuple, Union
 
 from async_upnp_client import UpnpRequester
 
 
-def read_file(filename) -> str:
+def read_file(filename: str) -> str:
     """Read file."""
     path = os.path.join("tests", "fixtures", filename)
     with open(path, "r") as file:
@@ -19,9 +20,18 @@ def read_file(filename) -> str:
 class UpnpTestRequester(UpnpRequester):
     """Test requester."""
 
-    def __init__(self, response_map) -> None:
+    def __init__(
+        self,
+        response_map: Mapping[
+            Tuple[str, str], Tuple[int, Mapping[Any, Any], Union[str, bytes, None]]
+        ],
+    ) -> None:
         """Class initializer."""
-        self._response_map = copy(response_map)
+        self.response_map: MutableMapping[
+            Tuple[str, str],
+            Tuple[int, MutableMapping[Any, Any], Union[str, bytes, None]],
+        ] = deepcopy(response_map)
+        self.exceptions: Deque[Optional[Exception]] = deque()
 
     async def async_do_http_request(
         self,
@@ -35,11 +45,16 @@ class UpnpTestRequester(UpnpRequester):
         # pylint: disable=too-many-arguments
         await asyncio.sleep(0.01)
 
-        key = (method, url)
-        if key not in self._response_map:
-            raise KeyError("Request not in response map")
+        if self.exceptions:
+            exception = self.exceptions.popleft()
+            if exception is not None:
+                raise exception
 
-        return self._response_map[key]
+        key = (method, url)
+        if key not in self.response_map:
+            raise KeyError(f"Request not in response map: {key}")
+
+        return self.response_map[key]
 
 
 RESPONSE_MAP = {
@@ -59,9 +74,19 @@ RESPONSE_MAP = {
         {"sid": "uuid:dummy", "timeout": "Second-300"},
         "",
     ),
+    ("SUBSCRIBE", "http://localhost:1234/upnp/event/AVTransport1"): (
+        200,
+        {"sid": "uuid:dummy-avt1", "timeout": "Second-150"},
+        "",
+    ),
     ("UNSUBSCRIBE", "http://localhost:1234/upnp/event/RenderingControl1"): (
         200,
         {"sid": "uuid:dummy"},
+        "",
+    ),
+    ("UNSUBSCRIBE", "http://localhost:1234/upnp/event/AVTransport1"): (
+        200,
+        {"sid": "uuid:dummy-avt1"},
         "",
     ),
 }