+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changes
 - Defer event callback URL determination until event subscriptions are created (@chishm)
 - Add `UpnpDevice.icons` and `UpnpProfileDevice.icon` to get URLs to device icons (@chishm)
 - Add more non-strict parsing of action responses (#68)
+- asyncio and aiohttp exceptions are wrapped in exceptions derived from `UpnpError` to hide implementation details and make catching easier (@chishm)
+- `UpnpProfileDevice` can resubscribe to services automatically, using an asyncio task (@chishm)
 
 
 0.18.0 (2021-05-23)